requirements.txt
## Common Packages
## ###############

Django==1.8.17
bcrypt==3.1.1
bootstrap-admin==0.3.6
dj-database-url==0.4.1
django-allauth==0.29.0
django-extensions==1.7.5
django-getenv==1.3.1
django-pipeline==1.5.4
django-pipeline-compass==0.1.5
django-rest-swagger==2.1.0
djangorestframework==3.5.3
drf-extensions==0.3.1
envdir==0.7
psycopg2==2.6.2
python-dateutil==2.6.0


## Production-only Packages
## ########################

gunicorn==19.6.0
whitenoise==3.2.2


## Development-only Packages
## TODO: Figure out how to install in development mode only.
## #########################################################

# bpython
# django-debug-toolbar
flake8==3.2.1
# pytest
# pytest-cov
# pytest-django
# pytest-xdist
Werkzeug==0.11.11
